Install zchunk before running tests in CI
We have a dependency on `unzck`. This is declared in the rpm spec file,
but not anywhere that CI can see it when running tests (on Ubuntu).
Some sort of container for running the tests would make sense, but for
the simplest fix, manually install it in CI  before running the tests.